empty_seat/requirements.txt
uvicorn[standard]
ultralytics
torch
pillow
python-multipart
opencv-python-headless
fastapi
numpy
